requirements.txt
discord.py==2.4.0
Flask==3.0.3
ics==0.7.2
pandas==2.2.2
pymongo==4.8.0
python-dotenv==1.0.1
